Inclusion of doxygen documentation in html format, can be viewed in "html/index.html".
Inclusion of "Doxygen" a config file for generation of doxygen documentation.

Inclusion of a vector struct "VectorPoint", to make it easy the management of the points inside some Line object. The documentation is almost complete, but some revision is necessary.

Inclusion of a Line object. Some features can be implemented like a function for remotion of points in the Line object. But it already covers almost all the requirements of this project. The goal is to make Line, a child object to a Rectangle object that will contain 4 Line objects. Also, the object Enemy Sight should be reimplemented with the inclusion of this new Line object.

The inclusion of the Line object made possible the reimplementation of "rotate45()" function. Now the signature is more intuitive. It receives a center point which is the reference for the rotation of a Line object.