Improvements in Line's function performace.
geometry.h was renamed to line.h. Now the geometric shapes will be in individual files.

A new function to return a pointer to a Point struct inside VectorPoint, was added. This made possible, to the functions of line.h, transfer parameters of Point type, as pointers.

Progress in the documentation of line.h and pointvector.h.