Enemy Completed.
Creation of a "log" module to print relevant information related to the game in a log file.

"personages" module separated for a module to each personage. This was done, because "enemy.c" was getting to big.

Enemy object completed and also functions in the "transformation" module was done to manipulate this object.

New enum type "Translation" added. Used to specify the translation direction for objects.